fix: align occupied_mask type with DEFINE_CHAIN_OPS MASK_TYPE
hook_chain_rw_t: uint8_t → uint16_t (matches macro arg)
fp_hook_chain_rw_t: uint16_t → uint32_t (matches macro arg)
Fix misleading slot count comments.

+++ include/hook.h
@@ -264,7 +264,7 @@ typedef struct hook_chain_rw
     hook_chain_rox_t *rox;
     int32_t chain_items_max;
     int32_t argno;
-    uint8_t occupied_mask;
+    uint16_t occupied_mask;
     int32_t sorted_indices[HOOK_CHAIN_NUM];
     int32_t sorted_count;
     hook_chain_item_t items[HOOK_CHAIN_NUM];
@@ -295,7 +295,7 @@ typedef struct fp_hook_chain_rw
     fp_hook_chain_rox_t *rox;
     int32_t chain_items_max;
     int32_t argno;
-    uint16_t occupied_mask;
+    uint32_t occupied_mask;
     int32_t sorted_indices[FP_HOOK_CHAIN_NUM];
     int32_t sorted_count;
     hook_chain_item_t items[FP_HOOK_CHAIN_NUM];
